Adjust component stylings
